sem4cepproject/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor>=3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email-validator]==2.9.2
python-dotenv==1.0.1
bcrypt==4.0.1
